+++ app/adapters/redis.py
@@ -1,9 +1,7 @@
 import redis
 
-from app.connectors.base import BaseConnector
 
-
-class RedisConnector(BaseConnector):
+class RedisAdapter:
     def __init__(self, host, port, user=None, password=None):
         self.password = password
         self.user = user
@@ -12,14 +10,13 @@ class RedisConnector(BaseConnector):
         self.redis_connection = None
 
     def connect(self):
-        # TODO need to try and reconnect in case of error
         self.redis_connection = redis.Redis(decode_responses=True)
 
     def disconnect(self):
         self.redis_connection.close()
 
-    def consumer(self, key: str) -> int | None:
+    def get_message_by_key(self, key: str) -> int | None:
         return self.redis_connection.get(key)
 
-    def producer(self, key: str, value: int) -> None:
+    def send_message_by_key(self, key: str, value: int) -> None:
         self.redis_connection.set(key, value)
