+++ app/main.py
@@ -1,9 +1,8 @@
+from adapters.rabbit import RabbitAdapter
 from config import settings
-from connectors.rabbit import RabbitConnector
-
 
 if __name__ == '__main__':
-    rmq_conn = RabbitConnector(
+    rmq_conn = RabbitAdapter(
         host=settings.rabbitmq_host,
         port=settings.rabbitmq_port,
         user=settings.rabbitmq_user,
