+++ app/repository/base.py
@@ -4,8 +4,6 @@ from typing import Any
 from sqlalchemy import select
 from sqlalchemy.orm import joinedload
 
-from app.exceptions import NotFoundException
-
 
 class BaseRepository(ABC):
     @abstractmethod
@@ -17,10 +15,6 @@ class BaseRepository(ABC):
         raise NotImplementedError
 
     async def delete(self, obj: Any):
-        raise NotImplementedError
-
-    @abstractmethod
-    async def get_or_error(self, instance_id: int, joined_load: tuple[str, ...] | None = None) -> Any | None:
         raise NotImplementedError
 
 
@@ -38,21 +32,3 @@ class BaseSQLAlchemyRepository(BaseRepository):
         if joined_load:
             base_q = base_q.options(*(joinedload(getattr(self.model_cls, field)) for field in joined_load))
         return (self.session.execute(base_q)).scalars().first()
-
-    async def delete(self, obj: Any):
-        return await self.session.delete(obj)
-
-    async def get_or_error(self, instance_id: int, joined_load: tuple[str, ...] | None = None) -> Any:
-        if instance := await self.get('id', instance_id, joined_load=joined_load):
-            return instance
-        raise NotFoundException(f'{self.model_cls.__name__} with id {instance_id} not found')
-
-    async def get_or_error_by_field(
-        self,
-        field: str,
-        value: Any,
-        joined_load: tuple[str, ...] | None = None,
-    ) -> Any:
-        if instance := await self.get(field, value, joined_load=joined_load):
-            return instance
-        raise NotFoundException(f'{self.model_cls.__name__} with {field} = {value} not found')
