+++ app/services/historical_transactions.py
@@ -1,10 +1,11 @@
 import json
 
+from app.config import logger
 from app.domain import HistoricalTransactions
 from app.services.uow import SqlAlchemyUnitOfWork
 
 
-def create_transaction(uow: SqlAlchemyUnitOfWork, data: str) -> dict | None:
+def create_transaction(uow: SqlAlchemyUnitOfWork, data: str) -> HistoricalTransactions | None:
     transaction_data = json.loads(data)
     with uow:
         method = uow.initial_data.get('id', transaction_data['id'])
@@ -16,12 +17,8 @@ def create_transaction(uow: SqlAlchemyUnitOfWork, data: str) -> dict | None:
         uow.historical_transactions.add(transaction)
         uow.commit()
         uow.refresh(transaction)
-        print(
+        logger.info(
             f'Created transaction {transaction.id} for {transaction.initial_data.provider_name} '
             f'with value {transaction.transaction_value}'
         )
-        return {
-            f'{transaction.initial_data.id}_{transaction.initial_data.provider_name}': int(
-                transaction.transaction_value
-            )
-        }
+        return transaction
